+++ pyiron_atomistics/dft/master/convergence_encut_parallel.py
@@ -3,7 +3,7 @@
 # Distributed under the terms of "New BSD License", see the LICENSE file.
 
 from __future__ import print_function
-from pyiron_atomistics.atomistics.master.parallel import AtomisticParallelMaster
+from pyiron_atomistics.master import AtomisticParallelMaster
 from pyiron_base import JobGenerator
 import numpy as np
