+++ pyiron_atomistics/generic/object_type.py
