+++ pyiron_atomistics/gpaw/pyiron_ase.py
@@ -3,7 +3,7 @@
 # Distributed under the terms of "New BSD License", see the LICENSE file.
 
 import numpy as np
-from pyiron_atomistics.atomistics.job.interactive import GenericInteractive
+from pyiron_atomistics.job import GenericInteractive
 from pyiron_atomistics.atomistics.structure.atoms import Atoms
 
 try:
