+++ pyiron_atomistics/interactive/quasi_newton.py
@@ -1,7 +1,7 @@
 import numpy as np
 from pyiron_base import GenericParameters
 import warnings
-from pyiron_atomistics.atomistics.job.interactivewrapper import (
+from pyiron_atomistics.job import (
     InteractiveWrapper,
     ReferenceJobOutput,
 )
