+++ pyiron_atomistics/job/atomistic.py
@@ -7,11 +7,11 @@ import warnings
 import numpy as np
 import numbers
 import os
-from typing import Callable, Union, Tuple
+from typing import Union, Tuple
 from ase.io import write as ase_write
 import posixpath
 from pyiron_atomistics.atomistics.structure.atoms import Atoms
-from pyiron_atomistics.atomistics.structure.neighbors import NeighborsTrajectory
+from pyiron_atomistics.structure.neighbors import NeighborsTrajectory
 from pyiron_atomistics.atomistics.structure.has_structure import HasStructure
 from pyiron_base import (
     GenericParameters,
