+++ pyiron_atomistics/job/sqs.py
@@ -6,7 +6,7 @@ import random
 import warnings
 import itertools
 from structuretoolkit.build import sqs_structures
-from pyiron_atomistics.atomistics.job.atomistic import AtomisticGenericJob
+from pyiron_atomistics.job.atomistic import AtomisticGenericJob
 from pyiron_base import state, DataContainer, GenericParameters, ImportAlarm
 from pyiron_atomistics.atomistics.structure.atoms import Atoms, ase_to_pyiron
 import numpy as np
