+++ pyiron_atomistics/master/murnaghan.py
@@ -11,7 +11,7 @@ from atomistics.evcurve.calculator import _strain_axes
 import matplotlib.pyplot as plt
 import numpy as np
 
-from pyiron_atomistics.atomistics.master.parallel import AtomisticParallelMaster
+from pyiron_atomistics.master.parallel import AtomisticParallelMaster
 from pyiron_atomistics.atomistics.structure.atoms import Atoms, ase_to_pyiron
 from pyiron_base import JobGenerator
