+++ pyiron_atomistics/master/sqsmaster.py
@@ -3,7 +3,7 @@
 # Distributed under the terms of "New BSD License", see the LICENSE file.
 
 import numpy as np
-from pyiron_atomistics.atomistics.master.parallel import AtomisticParallelMaster
+from pyiron_atomistics.master.parallel import AtomisticParallelMaster
 from pyiron_base import JobGenerator
 
 __author__ = "Jan Janssen"
