+++ pyiron_atomistics/structure/atom.py
