+++ pyiron_atomistics/structure/factories/aimsgb.py
