+++ pyiron_atomistics/structure/factories/compound.py
