+++ pyiron_atomistics/structure/structurestorage.py
@@ -16,7 +16,7 @@ from pyiron_base import FlattenedStorage, ImportAlarm
 from pyiron_atomistics.atomistics.structure.atom import Atom
 from pyiron_atomistics.atomistics.structure.atoms import Atoms
 from structuretoolkit.common.error import SymmetryError
-from pyiron_atomistics.atomistics.structure.neighbors import NeighborsTrajectory
+from pyiron_atomistics.structure.neighbors import NeighborsTrajectory
 import pyiron_atomistics.atomistics.structure.has_structure as pa_has_structure
 
 with ImportAlarm(
